Create NUM_WORD
We done this code in PYTHON language. We First take a function which converts the give string into LIST. And then, for each item in the list we convert the respected number into word using if-else statements. We can directly proceed with the string instead of converting it into list. After the function is created, we will give N as input and call our function.